+++ Sources/Notifly/notifly-ios-sdk/notifly-ios-sdk/SourceCodes/Util/Auth+Credentials.swift
